Improve event hierarchy for ContractDefinition
